some suggested improvements
- Centralised the Address type
- Centralised the formatDisplayAddress function
- isCapitalised should default to false as we want the default behaviour of the selector to stay the same
- use of shouldCapitalise was wrong, take a look at parameters and default parameters
