+++ src/components/Shared/FormElements/AddressFormComponents/SelectAddress.tsx
@@ -7,25 +7,16 @@ import DTESelect from "../../UI/DTESelect/DTESelect";
 import { Details, ContinueButton } from "./PostcodeLookup";
 import Utils from "../../../../Helper/Utils";
 import Honeypot from "../../Honeypot/Honeypot";
-
-type address = {
-  addressLine1: string;
-  addressLine2: string;
-  addressLine3: string;
-  addressLine4: string;
-  town: string;
-  postcode?: string;
-  fullAddress: string;
-};
+import { Address } from "../../../../types/ParticipantTypes";
 
 export type SelectAddressData = {
-  address: address;
+  address: Address;
   manualEntry?: boolean;
   changePostcode?: boolean;
 };
 
 interface SelectAddressProps {
-  addresses: address[];
+  addresses: Address[];
   postcode: string;
   nextButtonText?: string;
   hideInfo?: boolean;
@@ -128,7 +119,7 @@ const SelectAddress = (props: SelectAddressProps) => {
                 error={error?.message}
                 required={false}
                 options={[
-                  ...addresses.map((data: address, index: number) => {
+                  ...addresses.map((data: Address, index: number) => {
                     return {
                       value: index,
                       text: data.fullAddress,
@@ -136,6 +127,7 @@ const SelectAddress = (props: SelectAddressProps) => {
                   }),
                 ]}
                 onValueChange={onChange}
+                isCapitalised
               />
             </>
           )}
